Give it better shape to our postgresql integration.
Which means two things: properly separate the logic
of the writer module, so we can re-use all the bits there:
instrumentation, logic, cache-control, and so forth.

It also means we have a dummy implementation for storing
stat values in a simplistic table, which should be enough
before the final table format takes shape.